[server][cli] Update DBContextFactory to optionally use a given configuration and to include configuration from appsettings.json
The given configuration is expected to  be provided by the application host, which is normally responsible for generating the configuration object. When a configuration object isn't given (eg. when the db context is created while executing dotnet ef commands), one will be manually generated and now will include configuration from appsettings.json.